Splits Parser into BasicParser and Parser.
There is the need to decouple the concepts of parsing a sequence of PDF objects and parsing a PDF document.
The former process is part of the second one, and we need basic parsing features to be general and indipendent
enough to be used in various contexts (e.g. Content Streams). The parser before this commit was tied to the
cross-reference table to be able to work. In fact, it couldnt be used to parse a list of PDF objects in a
Content Stream. Now BasicParser can be used for that purpose.

TODO: test everything!